card component rename
