+++ apps/web/your-id-web/src/app/components/card-id/card-id.component.scss
